reuse EditSimaApplicationInput.tsx component for view
